# backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: 3rd Library API
  version: 1.0.0
  description: HTTP API for the backend.
servers:
  - url: /
...
paths:
  /api/users:
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUserRequest"
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/users/login:
    post:
      tags: [Users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserRequest"
      responses:
        "200":
          description: User logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/user:
    get:
      tags: [Users]
      summary: Get current user
      security:
        - TokenAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Users]
      summary: Update current user
      security:
        - TokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/user/pfp:
    post:
      tags: [Profile Pictures]
      summary: Upload profile picture
      security:
        - TokenAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Profile picture uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PfpResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/pfps/{pfpId}:
    get:
      tags: [Profile Pictures]
      summary: Download profile picture
      parameters:
        - $ref: "#/components/parameters/PfpId"
      responses:
        "200":
          description: Profile picture bytes
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Use format: Token <jwt>"
  parameters:
...
    PfpId:
      name: pfpId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Unauthorized
      headers:
        WWW-Authenticate:
          schema:
            type: string
          description: Token
...
    NotFound:
      description: Not found
    UnprocessableEntity:
      description: Unprocessable entity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorsResponse"
  schemas:
    ErrorsResponse:
      type: object
      required: [errors]
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [email, token, username, bio, image]
      properties:
        email:
          type: string
          format: email
        token:
          type: string
        username:
          type: string
        bio:
          type: string
        image:
          type: string
          nullable: true
          description: profile picture id
          format: uuid
    NewUserRequest:
      type: object
      required: [user]
      properties:
        user:
          type: object
          required: [username, email, password]
          properties:
            username:
              type: string
            email:
              type: string
              format: email
            password:
              type: string
    LoginUserRequest:
      type: object
      required: [user]
      properties:
        user:
          type: object
          required: [email, password]
          properties:
            email:
              type: string
              format: email
            password:
              type: string
    UpdateUserRequest:
      type: object
      required: [user]
      properties:
        user:
          type: object
          properties:
            email:
              type: string
              format: email
            username:
              type: string
            password:
              type: string
            bio:
              type: string
            image:
              type: string
              nullable: true
              description: profile picture id
              format: uuid
...
    PfpResponse:
      type: object
      required: [pfp]
      properties:
        pfp:
          $ref: "#/components/schemas/Pfp"
    Pfp:
      type: object
      required: [pfpId]
      properties:
        pfpId:
          type: string
          format: uuid
...